+++ classes/Gems/Snippets/Mail/MailModelFormSnippet.php
@@ -18,8 +18,13 @@
  * @license    New BSD License
  * @since      Class available since version 1.6.2
  */
-class Gems_Snippets_Mail_MailModelFormSnippet extends \Gems_Snippets_ModelFormSnippetGeneric
+class Gems_Snippets_Mail_MailModelFormSnippet extends \Gems_Snippets_ModelFormSnippetAbstract
 {
+    /**
+     * @var string The current selected language tab
+     */
+    protected $_currentLanguage;
+    
     /**
      *
      * @var \Gems_User_User
@@ -33,6 +38,11 @@ class Gems_Snippets_Mail_MailModelFormSnippet extends \Gems_Snippets_ModelFormSn
     protected $loader;
 
     /**
+     * @var \Zend_Locale
+     */
+    protected $locale;
+    
+    /**
      *
      * @var \Gems_Mail_MailElements
      */
@@ -43,6 +53,12 @@ class Gems_Snippets_Mail_MailModelFormSnippet extends \Gems_Snippets_ModelFormSn
      * @var \Gems_Mail_MailerAbstract
      */
     protected $mailer;
+
+    /**
+     *
+     * @var \MUtil_Model_ModelAbstract
+     */
+    protected $model;
 
     /**
      *
@@ -62,6 +78,7 @@ class Gems_Snippets_Mail_MailModelFormSnippet extends \Gems_Snippets_ModelFormSn
      */
     protected $view;
 
+    
     /**
      * Adds elements from the model to the bridge that creates the form.
      *
@@ -78,8 +95,10 @@ class Gems_Snippets_Mail_MailModelFormSnippet extends \Gems_Snippets_ModelFormSn
 
         $bridge->getForm()->getElement('gct_target')->setAttrib('onchange', 'this.form.submit()');
 
-        $defaultTab = $this->project->getLocaleDefault();
-        $this->mailElements->addFormTabs($bridge, 'gctt', 'active', $defaultTab, 'tabcolumn', 'gctt_lang', 'selectedTabElement', 'send_language');
+        $this->mailElements->addFormTabs($bridge, 'gctt', 
+                                         'active', $this->_currentLanguage, 
+                                         'tabcolumn', 'gctt_lang', 
+                                         'selectedTabElement', 'send_language');
 
         $config = array(
         'extraPlugins' => 'bbcode,availablefields',
@@ -95,13 +114,9 @@ class Gems_Snippets_Mail_MailModelFormSnippet extends \Gems_Snippets_ModelFormSn
             )
         );
 
-
-        $mailfields = $this->mailer->getMailFields();
-        foreach($mailfields as $field => $value) {
-            $mailfields[$field] = utf8_encode($value);
-        }
+        $mailfields = array_map('utf8_encode', $this->mailer->getMailFields());
+        ksort($mailfields);
         $config['availablefields'] = $mailfields;
-
 
         $config['availablefieldsLabel'] = $this->_('Fields');
         $this->view->inlineScript()->prependScript("
@@ -119,8 +134,9 @@ class Gems_Snippets_Mail_MailModelFormSnippet extends \Gems_Snippets_ModelFormSn
 
         $bridge->addElement($this->getSaveButton($bridge->getForm()));
 
-        $bridge->addElement($this->mailElements->createPreviewHtmlElement($this->_('Preview HTML')));
-        $bridge->addElement($this->mailElements->createPreviewTextElement($this->_('Preview Text')));
+        // $bridge->addElement($this->mailElements->createPreviewHtmlElement($this->_('Preview HTML')));
+        // $bridge->addElement($this->mailElements->createPreviewTextElement($this->_('Preview Text')));
+        
         $bridge->addHtml('available_fields', array('label' => $this->_('Available fields')));
     }
 
@@ -138,13 +154,52 @@ class Gems_Snippets_Mail_MailModelFormSnippet extends \Gems_Snippets_ModelFormSn
         parent::afterRegistry();
     }
 
+    /**
+     * @param string $language 
+     * @return mixed|null The language if it exists in the template, otherwise null 
+     */
+    protected function checkLanguage($language)
+    {
+        foreach($this->formData['gctt'] as $languageId => $templateLanguage) {
+            // Find the current template (for multi language projects)
+            if ($templateLanguage['gctt_lang'] == $language) {
+                return $language;
+            }
+        }
+    }
+
+    /**
+     * Creates the model
+     *
+     * @return \MUtil_Model_ModelAbstract
+     */
+    protected function createModel()
+    {
+        if ($this->model instanceof \Gems_Model_CommtemplateModel) 
+        {
+            $sub = $this->model->get('gctt', 'model');
+            
+            if ($sub instanceof \MUtil_Model_ModelAbstract && (! $sub->has('preview_html'))) {
+                $sub->set('preview_html', 'label', $this->_('Preview HTML'),
+                    'elementClass', 'Html',
+                    'noHidden', true
+                    );
+//                $sub->set('preview_text', 'label', $this->_('Preview Text'),
+//                          'elementClass', 'Html',
+//                          'noHidden', true
+//                );
+            }
+        }
+        
+        return $this->model;
+    }
 
     /**
      * Style the template previews
      * @param  array $templateArray template data
      * @return array html and text views
      */
-    protected function getPreview($templateArray)
+    protected function getPreview(&$templateArray)
     {
         $multi = false;
         if (count($templateArray) > 1) {
@@ -155,14 +210,19 @@ class Gems_Snippets_Mail_MailModelFormSnippet extends \Gems_Snippets_ModelFormSn
         $htmlView = \MUtil_Html::create()->div();
         $textView = \MUtil_Html::create()->div();
 
-        foreach($templateArray as $template) {
+        foreach($templateArray as &$template) {
             $content = '';
             if ($template['gctt_subject'] || $template['gctt_body']) {
                 if ($multi) {
                     $htmlView->h3()->append($allLanguages[$template['gctt_lang']]);
                     $textView->h3()->append($allLanguages[$template['gctt_lang']]);
                 }
-
+                
+                // Set the translate adapter to the local language
+                $locale = new \Zend_Locale($template['gctt_lang']);
+                $this->translateAdapter->setLocale($locale);
+                $this->mailer->setLanguage($template['gctt_lang'], true);
+                
                 $content .= '[b]';
                 $content .= $this->_('Subject:');
                 $content .= '[/b] [i]';
@@ -170,14 +230,21 @@ class Gems_Snippets_Mail_MailModelFormSnippet extends \Gems_Snippets_ModelFormSn
                 $content .= "[/i]\n\n";
                 $content .= $this->mailer->applyFields($template['gctt_body']);
 
-                $htmlView->div(array('class' => 'mailpreview'))
-                        ->raw(\MUtil_Markup::render($content, 'Bbcode', 'Html'));
-                $textView->pre(array('class' => 'mailpreview'))
-                        ->raw(wordwrap(\MUtil_Markup::render($content, 'Bbcode', 'Text'), 80));
-            }
+                $template['preview_html'] = \MUtil_Html::create()->div(['class' => 'mailpreview']);
+                $template['preview_html']->raw(\MUtil_Markup::render($content, 'Bbcode', 'Html'));
+                $template['preview_text'] = \MUtil_Html::create()->pre(['class' => 'mailpreview']);
+                $template['preview_text']->raw(wordwrap(\MUtil_Markup::render($content, 'Bbcode', 'Text'), 80));
 
+                $htmlView->append($template['preview_html']);
+                $textView->append($template['preview_text']);
+                
+            }
         }
 
+        // Reset fields to the current language 
+        $this->translateAdapter->setLocale($this->locale);
+        $this->mailer->setLanguage($this->locale->getLanguage(), true);
+        
         return array('html' => $htmlView, 'text' => $textView);
     }
 
@@ -196,6 +263,8 @@ class Gems_Snippets_Mail_MailModelFormSnippet extends \Gems_Snippets_ModelFormSn
     {
         parent::loadFormData();
         $this->loadMailer();
+
+        $this->formData['available_fields'] = $this->mailElements->displayMailFields($this->mailer->getMailFields());
 
         if (isset($this->formData['gctt'])) {
             $multi = false;
@@ -219,7 +288,15 @@ class Gems_Snippets_Mail_MailModelFormSnippet extends \Gems_Snippets_ModelFormSn
             }
         }
 
-        $this->formData['available_fields'] = $this->mailElements->displayMailFields($this->mailer->getMailFields());
+        if (! $this->_currentLanguage && isset($this->formData['send_language'])) {
+            $this->_currentLanguage = $this->checkLanguage($this->formData['send_language']);
+        }
+        if (! $this->_currentLanguage) {
+            $this->_currentLanguage = $this->checkLanguage($this->locale->getLanguage());
+            if (! $this->_currentLanguage) {
+                $this->_currentLanguage = $this->project->getLocaleDefault();
+            }
+        }
     }
 
     /**
@@ -247,21 +324,21 @@ class Gems_Snippets_Mail_MailModelFormSnippet extends \Gems_Snippets_ModelFormSn
 
             if (! empty($this->formData['preview'])) {
                 $this->addMessage($this->_('Preview updated'));
+                \MUtil_Echo::track(array_keys($this->formData), $this->formData['select_template']);
                 return;
             }
 
             if (! empty($this->formData['sendtest'])) {
+                // \MUtil_Echo::track(array_keys($this->formData['gctt'][1]), $this->formData['send_language']);
                 $this->mailer->setTo($this->formData['to']);
 
                 // Make sure at least one template is set (for single language projects)
                 $template   = reset($this->formData['gctt']);
                 $languageId = key($this->formData['gctt']);
-                if ($this->formData['send_language']) {
-                    foreach($this->formData['gctt'] as $languageId => $templateLanguage) {
-                        // Find the current template (for multi language projects)
-                        if ($templateLanguage['gctt_lang'] == $this->formData['send_language']) {
-                            $template = $templateLanguage;
-                        }
+                foreach($this->formData['gctt'] as $languageId => $templateLanguage) {
+                    // Find the current template (for multi language projects)
+                    if ($templateLanguage['gctt_lang'] == $this->_currentLanguage) {
+                        $template = $templateLanguage;
                     }
                 }